fix: prevent `Utils.keyStroke` to re-enter eventloop
This change will prevent `Utils.keyStroke` actions to be ran on os level
and will not re-enter the eventloop.

Which means that use will not be able to remap keys with mapped keys.
Say we have set key `t` for `cmd+t`, previously we can map `T` to `t`,
but now it does nothing. User have to explicitly set the native
keystrokes, in this case `T` with `cmd+t` to work.